Safer and more reliable device getter
The former getDevices function did not list anything at all for me, so I investigated and found out, that I can list out /proc/asound/pcm file to get all sound devices in a format that is way easier to parse. Also, I thing synchronicity in this case is preferred.